:white_check_mark: Split up tests
